backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
sqlalchemy[asyncio]==2.0.36
asyncpg==0.30.0
psycopg2-binary==2.9.10
alembic==1.14.0
pydantic==2.10.4
pydantic-settings==2.7.0
python-dotenv==1.0.1
PyYAML==6.0.2
cryptography==43.0.3
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
apscheduler==3.10.4
httpx==0.28.1
mcp>=1.0.0,<1.1.0
jinja2==3.1.4
tiktoken==0.8.0
